requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
black==22.8.0
certifi==2022.6.15.1
charset-normalizer==2.1.1
click==8.1.3
Django==3.2.15
django-mathfilters==1.0.0
djangorestframework==3.13.1
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.12.0
isort==5.10.1
mypy-extensions==0.4.3
pathspec==0.10.1
platformdirs==2.5.2
psycopg2-binary==2.8.6
python-dotenv==0.21.0
pytz==2022.2.1
requests==2.28.1
sqlparse==0.4.2
stripe==4.1.0
tomli==2.0.1
typed-ast==1.5.4
typing_extensions==4.3.0
urllib3==1.26.12
zipp==3.8.1
